# docs/swagger.yaml
components:
  schemas:
    dto.BookResponse:
      properties:
        author:
          type: string
        id:
          type: integer
        title:
          type: string
      type: object
    dto.CreateBookRequest:
      properties:
        author:
          type: string
        title:
          type: string
      required:
      - author
      - title
      type: object
    dto.ErrorResponse:
      properties:
        error:
          type: string
      type: object
    dto.Pagination:
      properties:
        current_page:
          type: integer
        last_page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
      type: object
    dto.PaginationResponse-dto_BookResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/dto.BookResponse'
          type: array
          uniqueItems: false
        pagination:
          $ref: '#/components/schemas/dto.Pagination'
      type: object
    dto.SuccessResponse-dto_BookResponse:
      properties:
        data:
          $ref: '#/components/schemas/dto.BookResponse'
      type: object
    dto.UpdateBookRequest:
      properties:
        author:
          type: string
        title:
          type: string
      type: object
  securitySchemes:
    Bearer:
      description: Bearer token authentication
      in: header
      name: Authorization
      type: apiKey
externalDocs:
  description: ""
  url: ""
info:
  title: GO-FIBER-TEMPLATE API
  version: "1.0"
openapi: 3.1.0
paths:
  /books:
    get:
      description: get books
      parameters:
      - description: Number of history to be retrieved
        in: query
        name: limit
        schema:
          type: integer
      - description: Page to retrieved
        in: query
        name: page
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.PaginationResponse-dto_BookResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Internal Server Error
      summary: GetBooks
      tags:
      - book
    post:
      description: create book by title and author
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.CreateBookRequest'
        description: Book Data
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_BookResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Internal Server Error
      summary: CreateBook
      tags:
      - book
  /books/{id}:
    delete:
      description: delete book by id
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Internal Server Error
      summary: DeleteBook
      tags:
      - book
    get:
      description: get book by id
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_BookResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Internal Server Error
      summary: GetBook
      tags:
      - book
    patch:
      description: update book data
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/dto.UpdateBookRequest'
        description: Book Data
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.SuccessResponse-dto_BookResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dto.ErrorResponse'
          description: Internal Server Error
      summary: UpdateBook
      tags:
      - book